+++ Block/Form/Field/AbstractField.php
@@ -163,7 +163,7 @@ class AbstractField extends \Magento\Framework\View\Element\Template implements 
      */
     public function getFieldClass()
     {
-        return $this->isRequired() ? 'required' : 'aaa';
+        return $this->isRequired() ? 'required' : '';
     }
 
     /**
@@ -180,7 +180,7 @@ class AbstractField extends \Magento\Framework\View\Element\Template implements 
      */
     public function getNote()
     {
-       return $this->getField()->getNote();
+        return $this->getField()->getFrontendNote();
     }
 
     /**